% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Antibiotics.R
\docType{data}
\name{Antibiotics}
\alias{Antibiotics}
\title{Example data set for antibiotic resistance data}
\format{A flat vector of size 500.}
\usage{
Antibiotics
}
\description{
This data set is a simulated data set with n = 500 observations.
}
\details{
This simulated data set may be used to test and work with the functionality of the "ECOFFBayes" package. In total it contains 500 observations
that should be similar to the EUCAST data sets.
}
\keyword{datasets}
