% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateNormal.R
\name{EstimateNormal}
\alias{EstimateNormal}
\title{Normal Components}
\usage{
EstimateNormal(gibbs, k, y, prior)
}
\arguments{
\item{gibbs}{Matrix with parameters sampled with gibbs algorithm}

\item{k}{Number components}

\item{y}{Data}

\item{prior}{Prior list}
}
\description{
Calculates the log likelihood for each step in the gibbs sampling.
}
