% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoff.R
\name{FindECOFF}
\alias{FindECOFF}
\title{ECOFF Finder}
\usage{
FindECOFF(obj, pi.level = 0.3, quantile = 0.01)
}
\arguments{
\item{obj}{object of class TheBayesteApproach}

\item{pi.level}{The level of summed up component probabilities from the distributions coming from the "right" side that needs to be exceeded (default is 0.3)}

\item{quantile}{The quantile from the component where the pi.level is exceed (default is 0.01)}
}
\description{
This function calculates the ECOFF for the wild type of an 'TheBayesteApproach' object
}
\details{
This function takes an object of the class TheBayesteApproach and calculates the ECOFF. The ECOFF is the
epidemiological cut-off value that separates bacteria without (wild type) and with acquired resistance
mechanisms (non-wild type) to the antibiotics in question. In order to caluclate the ECOFF, first the wild type
component is selected by finding the component where the sum of its \eqn{\pi_k} and  the \eqn{\pi}'s of all components
with a higher \eqn{\mu_k} exceed a certain threshold (by default: 0.3). Once this component is found a certain
quantile is taken as the ECOFF value (by default: 0.01)
}
\examples{

\donttest{data("Antibiotics")
bayes.density <- TheBayesteApproach(Antibiotics)
FindECOFF(bayes.density, pi.level = 0.3, quantile = 0.02)
plot(bayes.density, ECOFF = TRUE)}

}
\keyword{ECOFF,}
\keyword{antibiotics}
\keyword{type,}
\keyword{wild}
