% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNumberComponents.R
\name{GetNumberComponents}
\alias{GetNumberComponents}
\title{Number of Components}
\usage{
GetNumberComponents(gibbs)
}
\arguments{
\item{gibbs}{Matrix with parameters sampled (output of function CallJags)}
}
\value{
Number of components and MAPE estimates of the parameters of the normal components
}
\description{
This internal function postprocesses the gibbs sampling output and returns the resulting number of components and MAPE estimates
}
\keyword{Gibbs}
\keyword{sampler}
