% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateNormal.R
\name{LogLikelihood}
\alias{LogLikelihood}
\title{Normal Components}
\usage{
LogLikelihood(gibbs, k, y, prior)
}
\arguments{
\item{gibbs}{matrix with parameters sampled with gibbs algorithm}

\item{k}{Number of components}

\item{y}{vector of flat data}

\item{prior}{an optional list of hyperparameters for prior distributions}
}
\value{
logLikelihood
}
\description{
This internal function calculates the log likelihood for each step in the gibbs sampling
}
\keyword{logLikelihood}
