% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHist.R
\name{plot.TheBayesteApproach}
\alias{plot.TheBayesteApproach}
\title{Plot TheBayesteApproach}
\usage{
\method{plot}{TheBayesteApproach}(x, xlim = c(6, 50), xlab = "ZD",
  main = "Density estimator", ECOFF = FALSE, pi.level = 0.3,
  quantile = 0.01, ...)
}
\arguments{
\item{x}{An object of class TheBayesteApproach}

\item{xlim}{limits of x axis}

\item{xlab}{label of x axis}

\item{main}{the plot title}

\item{ECOFF}{logical value indicating if the ECOFF should be computed and added to the plot}

\item{pi.level}{The level of summed up pi from the distributions coming from the "right" side that needs to be exceeded, ECOFF argument (Default: 0.3)}

\item{quantile}{quantile of the component with the highest mean that is taken for the ECOFF. Default is 0.01}

\item{...}{additional plot parameters}
}
\description{
This function plots a histogram of the data and its estimated mixture density.
}
\details{
This function plots a histogram of the data and its estimated mixture density. Additionally to the plotted density,
an ECOFF value can be added in the plot by setting the ECOFF argument to true. Further, the arguments pi.level and
qunatile can be defined to calculate the ECOFF in a certain way (see \code{\link{FindECOFF}} for Details). Further
plot arguments may be provided.
}
\examples{
\donttest{data("Antibiotics")
bayes.density <- TheBayesteApproach(Antibiotics)
print(bayes.density)
plot(bayes.density)}
}
\keyword{antibiotics}
\keyword{density,}
\keyword{plot.TheBayesteApproach,}
\keyword{plotting}
